requirements.txt
django
djangorestframework
environs
django-enumchoicefield
numpy
